Renames the streaming steps to quickstart, moves it out from user route, makes the Notice Types object an export, filters out the parent notice types (ex. Fermi), typo fix
