+++ app/routes/index.mdx
@@ -31,7 +31,7 @@ import {
       GCN distributes alerts between space- and ground-based observatories,
       physics experiments, and thousands of astronomers around the world.
     </p>
-    <Link to="/user/streaming_steps" className="usa-button">
+    <Link to="/quickstart" className="usa-button">
       <>Start streaming GCN Notices</>
     </Link>
   </div>
@@ -89,7 +89,7 @@ There are three ways to stream GCN Notices in real time:
       <p>Three formats, one protocol.</p>
     </CardBody>
     <CardFooter>
-      <Link className="usa-button" to="/user/streaming_steps">
+      <Link className="usa-button" to="/quickstart">
         <>Get Started</>
       </Link>
     </CardFooter>
