+++ app/routes/quickstart/code.tsx
@@ -10,14 +10,18 @@ import Tabs from '~/components/Tabs'
 import { ClientSampleCode } from '~/components/ClientSampleCode'
 import { Form, useLoaderData } from '@remix-run/react'
 import type { DataFunctionArgs } from '@remix-run/node'
-import { ClientCredentialVendingMachine } from '../client_credentials.server'
+import { ClientCredentialVendingMachine } from '../user/client_credentials.server'
 import { Button } from '@trussworks/react-uswds'
+import { NoticeTypes } from '~/components/NoticeTypeCheckboxes'
 
 export async function loader({ request }: DataFunctionArgs) {
   const { clientId, noticeFormat, ...rest } = Object.fromEntries(
     new URL(request.url).searchParams
   )
-  const noticeTypes = Object.keys(rest)
+  const noticeTypes = Object.keys(rest).filter(
+    (noticeType) => Object.keys(NoticeTypes).indexOf(noticeType) == -1
+  )
+  console.log(noticeTypes)
   const machine = await ClientCredentialVendingMachine.create(request)
   const clientCredentialProps = await machine.getClientCredential(clientId)
   return {
@@ -50,7 +54,7 @@ export default function Code() {
       }),
     },
     {
-      label: 'Javscript',
+      label: 'Javascript',
       Component: ClientSampleCode({
         clientId,
         clientSecret,
